# server/spec/openapi.yaml
openapi: "3.1.0"
info:
  title: Prompt Mint API
  version: "1.0.0"
  description: |
    Prompt Mint marketplace API — Express backend serving prompt listing,
    user management, governance, appeals, licensing, webhooks, versioning,
    notifications, analytics, and API key management.
  contact:
    name: Prompt Mint Team
  license:
    name: MIT
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: https://api.promptmint.io
    description: Production server
paths:
...
  /api/webhooks:
    post:
      operationId: registerWebhook
      summary: Register a new webhook subscription
      tags: [Webhooks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  format: uri
                events:
                  type: array
                  items:
                    type: string
                secret:
                  type: string
      responses:
        "201":
          description: Webhook registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookSubscription"
    get:
      operationId: getWebhook
      summary: Get webhook subscription details
      tags: [Webhooks]
      responses:
        "200":
          description: Webhook subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookSubscription"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteWebhook
      summary: Delete a webhook subscription
      tags: [Webhooks]
      responses:
        "200":
          description: Webhook deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
...
  /api/webhooks/dead-letters/{id}/replay:
    post:
      operationId: replayWebhookDeadLetter
      summary: Replay a dead-letter webhook event
      tags: [Webhooks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Replay queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean

  # ─── Notifications ─────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    adminToken:
      type: http
      scheme: bearer
      description: Admin API token from ADMIN_API_TOKEN env var
    apiKeyQuery:
      type: apiKey
      in: header
      name: X-Api-Key
      description: API key for programmatic access
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: string
...
  schemas:
...
    WebhookSubscription:
      type: object
      properties:
        _id:
          type: string
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...